+++ src/applications/financial-status-report/pages/assets/monetary.js
