+++ src/applications/financial-status-report/pages/index.js
@@ -14,11 +14,11 @@ import * as spouseBenefits from './householdIncome/spouseBenefits';
 import * as spouseEmployment from './householdIncome/spouseEmployment';
 import * as spousePreviousEmployment from './householdIncome/spousePreviousEmployment';
 import * as dependents from './householdIncome/dependents';
-import * as monetary from './householdAssets/monetary';
-import * as realEstate from './householdAssets/realEstate';
-import * as recreationalVehicles from './householdAssets/recreationalVehicles';
-import * as vehicles from './householdAssets/vehicles';
-import * as otherAssets from './householdAssets/otherAssets';
+import * as monetary from './assets/monetary';
+import * as realEstate from './assets/realEstate';
+import * as recreationalVehicles from './assets/recreationalVehicles';
+import * as vehicles from './assets/vehicles';
+import * as otherAssets from './assets/otherAssets';
 import * as spouseInformation from './householdIncome/spouseInformation';
 import * as spouseSocialSecurity from './householdIncome/spouseSocialSecurity';
 import * as expenses from './householdExpenses/expenses';
@@ -27,7 +27,7 @@ import * as repayments from './householdExpenses/repayments';
 import * as otherExpenses from './householdExpenses/otherExpenses';
 import * as resolutionOptions from './resolutionOptions/resolutionOptions';
 import * as resolutionComments from './resolutionOptions/resolutionComments';
-import * as bankruptcyHistory from './bankruptcyAttestation/bankruptcyHistory';
+import * as bankruptcyHistory from './bankruptcy/history';
 import * as contactInfo from './veteranInformation/contactInformation';
 
 export {
